#50: Create test helpers.py. Add items. Update test imports
